Implementing the cmake-helpers
The implementation is similar to a BusyBox like implementation where
everything is just in one box that can do many things.
This looks at the name of the script as it's called to decide how the
program should behave. If it's called from cmake-cp, it will copy the
build directory, if it's called from cmake-mv, it will move the build
directory.

I'm also including updates to the readme to reflect the implementation.