Adding test suite
This replaces the sad little test I had before with some actual tests.

This commit tests two functions, which are fairly central to the
operation of the program.

The first three tests look at the behaviour of get_cache_name, which
will throw an error if the incorrect file is passed to it, or if it's
given a directory, if that directory doesn't contain a CMakeCache.txt
file. If it does, the function should find the file and return the full
absolute path to the CMakeCache.txt file.

The second function is the ChangeBaseDir test, which is what goes
through each line of the CMakeCache file and replaces the original build
path with the new build path and then moves it back. If the two match,
and the intermediate doesn't match, then it worked "correctly". This
could be a better test that actually pattern matches which lines should
be replaced and checks that they are fixed correctly, but it works for
now.