Splitting out dest. directory creation
I had duplicated the logic to create the destination directory. I've
split that out. Its not completely done, but does the basics.

Current Logic:

First, if we want to allow overwriting and the directory/file already
exists, delete it.

Next, if the directory doesn't exist (either we deleted it, or it never
existed), create a new directory.

What's missing? If overwriting is not allowed and the directory/file
name is already taken, then it will silently fail and do lots of bad
things later on in the program. This should probably raise an exception
of some sort and crash the program.